components/common/python/requirements.txt
requests==2.21.0
pytz==2019.2
python_dateutil==2.8.1
kombu==4.6.8
pebble==4.6.0